Fixed debug build assertions for llvm-8.0
